+++ modules/tinymce/src/core/main/ts/content/GetContentImpl.ts
@@ -1,5 +1,5 @@
-import { Arr, Fun, Optional, Type, Unicode } from '@ephox/katamari';
-import { Attribute, Css, Html, Insert, Remove, SelectorFilter, SugarElement, SugarShadowDom } from '@ephox/sugar';
+import { Fun, Optional, Type } from '@ephox/katamari';
+import { Attribute, Css, Html, Insert, Remove, SugarElement, SugarShadowDom } from '@ephox/sugar';
 
 import Editor from '../api/Editor';
 import AstNode from '../api/html/Node';
@@ -8,36 +8,13 @@ import Tools from '../api/util/Tools';
 import * as ElementType from '../dom/ElementType';
 import * as TrimHtml from '../dom/TrimHtml';
 import * as Zwsp from '../text/Zwsp';
+import { cleanupBogusElements, cleanupInputNames } from './ContentCleanup';
 import { Content, GetContentArgs } from './ContentTypes';
 
 const trimEmptyContents = (editor: Editor, html: string): string => {
   const blockName = Options.getForcedRootBlock(editor);
   const emptyRegExp = new RegExp(`^(<${blockName}[^>]*>(&nbsp;|&#160;|\\s|\u00a0|<br \\/>|)<\\/${blockName}>[\r\n]*|<br \\/>[\r\n]*)$`);
   return html.replace(emptyRegExp, '');
-};
-
-export const cleanupBogusElements = (parent: SugarElement<Node>): void => {
-  const bogusElements = SelectorFilter.descendants(parent, '[data-mce-bogus]');
-  Arr.each(bogusElements, (elem) => {
-    const bogusValue = Attribute.get(elem, 'data-mce-bogus');
-    if (bogusValue === 'all') {
-      Remove.remove(elem);
-    } else if (ElementType.isBr(elem)) {
-      // Need to keep bogus padding brs represented as a zero-width space so that they aren't collapsed by the browser
-      Insert.before(elem, SugarElement.fromText(Unicode.zeroWidth));
-      Remove.remove(elem);
-    } else {
-      Remove.unwrap(elem);
-    }
-  });
-};
-
-export const cleanupInputNames = (parent: SugarElement<Node>): void => {
-  const inputs = SelectorFilter.descendants(parent, 'input');
-
-  Arr.each(inputs, (input) => {
-    Attribute.remove(input, 'name');
-  });
 };
 
 const getPlainTextContent = (editor: Editor, body: HTMLElement) => {