+++ modules/tinymce/src/core/test/ts/browser/content/EditorGetContentTextFormatTest.ts
@@ -11,19 +11,14 @@ describe('browser.tinymce.core.content.EditorGetContentTextFormatTest', () => {
   }, []);
 
   const assertSelectedRadioButtons = (editor: Editor, nrOfInputs: number, shouldBeSelected: number) => {
-    const inputs = editor.getBody().getElementsByTagName('input');
-    let selected = 0;
-    for (let index = 0; index < inputs.length; index++) {
-      if (inputs[index].checked) {
-        selected++;
-      }
-    }
+    const total = editor.getBody().querySelectorAll('input').length;
+    const selected = editor.getBody().querySelectorAll('input:checked').length;
 
-    assert.equal(inputs.length, nrOfInputs, 'Should have the right amount of inputs');
+    assert.equal(total, nrOfInputs, 'Should have the right amount of inputs');
     assert.equal(selected, shouldBeSelected, 'Should have exactly one radio button');
   };
 
-  it('TDA: get text format content should trim zwsp', () => {
+  it('TBA: get text format content should trim zwsp', () => {
     const editor = hook.editor();
     editor.setContent('<p>' + Zwsp.ZWSP + 'a</p>');
     const html = editor.getContent({ format: 'text' });
@@ -32,8 +27,7 @@ describe('browser.tinymce.core.content.EditorGetContentTextFormatTest', () => {
 
   it('TINY-7981: inputs should not be deselected', () => {
     const editor = hook.editor();
-    // eslint-disable-next-line max-len
-    editor.setContent('<p><label><input name="group-name" type="radio" value="1">Option 1</label><label><input checked="checked" name="group-name" type="radio" value="2">Option 2</label><label><input name="group-name" type="radio" value="3">Option 3</label></p>');
+    editor.setContent('<p><input name="group-name" type="radio">Option 1<input checked="checked" name="group-name" type="radio">Option 2<input name="group-name" type="radio">Option 3</p>');
     const html = editor.getContent({ format: 'text' });
     assert.equal(html, 'Option 1Option 2Option 3', 'Should be expected text');
     assertSelectedRadioButtons(editor, 3, 1);